#33 wip remove shared api directory
